Fixed bug w/ system_python_path() that caused an infinite loop.
system_python_path() would find the next `python` in `$PATH`, excluding itself.
If `pyenv` is installed, then this would be `$PYENV_ROOT/shims/python`,
which would then try to call the first `python` in `$PATH` as the system python.
Except this would be ourself, i.e., `$CARGO_HOME/bin/python`,
creating an infinite loop back and forth.

This is resolved by skipping `$PYENV_ROOT/shims/python` if it exists,
just like current_exe() is already skipped.

Also, so that `python-path` works for this, too,
the current_exe is always assumed to have a file name of `python`.