Release 0.2.0 (from 0.1.1):
Added support for symlinks to `python` to invoke python scripts w/o `pyenv`'s overhead.
`pyenv` installs shims to all installed scripts that invoke `pyenv` to run the script.
`pyenv` is slow, so this can add up to 1 second to the script startup.
If my `python` wrapper is symlinked to `$script`,
then running `./$script` runs `python $(dirname $(which python))/$script`,
where this `python` is the real `python` here.
If `$script` is actually not a script, it's just run instead of `python`,
and the same is done if it's a non-Python script.

Also, removed `python-path` executable and replaced it with `python --path`.
Also added `python --prefix`, which is the same as `dirname $(python --path)`.
These aren't in normal `python`, but they don't clash with any `python` commands.