Improved the library (lib.rs mainly) with better APIs (errors especially). For example, where the pyenv python was found (like shell, local, global) is now tracked, too.
The correct path, etc. are printed for `python --path`, etc. when using symlinks to scripts,
instead of the python path previously for scripts.

Added `python --which`, which shows which python (or script/binary) is being run and where and why it was found.

Updated benchmarks.

Removed the panic abort and size optimization in profile.release.
Not really important and stack traces are nicer.